requirements.txt
fastapi
uvicorn
pandas
openpyxl
sentence-transformers
faiss-cpu
python-dotenv
google-generativeai
jinja2
